+++ xyz.struthers.rhul.ham/src/main/java/xyz/struthers/rhul/ham/config/Properties.java
@@ -31,7 +31,7 @@ public class Properties {
 	// "D:\\Git\\compham\\xyz.struthers.rhul.ham"; // NUC
 	public static final String RESOURCE_DIRECTORY = HOME_DIRECTORY + "\\src\\main\\resources";
 
-	public static final double SUPERANNUATION_RATE = 0.095; // 9.5%
+	public static final float SUPERANNUATION_RATE = 0.095f; // 9.5%
 	public static final long RANDOM_SEED = 20190315L;
 
 	// static data fields
